Add WebSocket KISS remote, runtime LoRa sync word control, fix promiscuous-mode packet loss
- Add WebSocketRemote.h: WS KISS listener (port 7634), off by default,
  mutually exclusive with the TCP wifi-remote host (shared KISS parser
  state can't take two live hosts at once)
- Add CMD_SYNC_WORD to get/set the LoRa sync word at runtime across
  sx126x/sx127x/sx128x; session-only (never persisted), always reverts
  to the Reticulum default on radio restart
- Fix promiscuous-mode packets on ESP32/nRF52 never reaching the host:
  they set an unrelated, unread packet_ready global instead of the local
  ready flag that actually gates enqueueing to modem_packet_queue
- Add missing sx127x::reset() (sx126x/sx128x already had it;
  validate_status() calls it uniformly on any modem when no radio is found)
- Fix bt_ssp_pin getting stuck non-zero when pairing is interrupted before
  bt_disable_pairing()/bt_pairing_complete(), which permanently hid the
  external framebuffer; gate on bt_state == BT_STATE_PAIRING instead
- Fix temperature reporting truncating instead of rounding
- Fix T096 battery voltage label ("d.dd" -> "d.dV")

+++ sx128x.h
@@ -68,7 +68,7 @@ public:
   void setCodingRate4(int denominator);
   uint8_t getCodingRate4();
   void setPreambleLength(long preamble_symbols);
-  void setSyncWord(int sw);
+  void setSyncWord(uint8_t sw);
   bool dcd();
   void clearIRQStatus();
   void enableCrc();
